Add styling and correct hooks render synchronicity in frontend
