+++ frontend/src/Collection/GameCardCollection.tsx
@@ -1,16 +1,19 @@
 import {GameCard} from "../Game/GameCard.ts";
 import axios from "axios";
-import {useEffect, useState} from "react";
+import React, {useEffect, useState} from "react";
 import GameCardFrame from "./GameCardFrame.tsx";
-import {IconButton, Modal} from "@mui/material";
+import {Button, Card, CardActions, CardContent, IconButton, Modal, Stack, TextField, Typography} from "@mui/material";
 import {AddCircle} from "@mui/icons-material";
-import NewGameCard from "./NewGameCard.tsx";
+import CheckBoxIcon from "@mui/icons-material/CheckBox";
+import DoDisturbOnIcon from "@mui/icons-material/DoDisturbOn";
 import "./GameCardCollection.css"
 
 export default function GameCardCollection() {
 
     const [allNonDefaultGameCards, setAllNonDefaultGameCards] = useState<GameCard[]>([]);
-    const [, setAllCardSetNames] = useState<string[]>([])
+    const [, setAllCardSetNames] = useState<string[]>([]);
+    const [title, setTitle] = useState<string>("");
+    const [cardSetName, setCardSetName] = useState<string>("");
     const [isModalOpen, setIsModalOpen] = useState(false);
 
     useEffect(
@@ -42,11 +45,35 @@ export default function GameCardCollection() {
         setAllCardSetNames(Array.from(uniqueSetNames))
     }
 
+    function saveGameCard() {
+        axios.post(
+            "/api/game_cards", {
+                "title": title,
+                "cardSetName": cardSetName
+            } as GameCard)
+            .then(() => {
+                setTitle("")
+                setCardSetName("")
+                loadAllNonDefaultGameCards()
+            })
+            .catch(console.error)
+            .finally(closeModal)
+    }
+
+    function inputTitle(event: React.ChangeEvent<HTMLInputElement>) {
+        setTitle(event.target.value)
+    }
+
+    function inputCardSetName(event: React.ChangeEvent<HTMLInputElement>) {
+        setCardSetName(event.target.value)
+    }
+
     const openModal = () =>
         setIsModalOpen(true);
 
-    const closeModal = () =>
-        setIsModalOpen(false);
+    const closeModal = () => {
+        setIsModalOpen(false)
+    };
 
 
     return (
@@ -64,9 +91,40 @@ export default function GameCardCollection() {
                 open={isModalOpen}
                 sx={{mt: 20, ml: 6}}
             >
-                <NewGameCard onClose={closeModal} loadCards={loadAllNonDefaultGameCards}/>
+                <Card className="new-card" sx={{
+                    maxWidth: 220,
+                    margin: 3,
+                    background: "#FDF6E1",
+                    boxShadow: 0,
+                    border: 0.5,
+                    borderColor: "rgba(122,119,119,0.3)",
+                    borderRadius: '15px'
+                }}>
+                    <CardContent>
+                        <Typography sx={{fontSize: 14}} color="text.secondary" gutterBottom>
+                            Create new game card
+                        </Typography>
+                        <Typography variant="h5" component="div">
+                            <TextField id="title" label="card title?" onInput={inputTitle}/>
+                        </Typography>
+                        <Typography sx={{mb: 1.5}} color="text.secondary">
+                            <TextField id="set" label="set name?" onInput={inputCardSetName}/>
+                        </Typography>
+                    </CardContent>
+                    <CardActions>
+                        <Stack className={"new-card-stack"} direction="row" paddingBottom={3}>
+                            <Button id={"new-card-button"} onClick={saveGameCard} sx={{
+                                m: 5, maxWidth: 60, color: "#508356", boxShadow: 0, borderRadius: '15px'
+                            }}><CheckBoxIcon/></Button>
+                            <Button id={"new-card-button"} onClick={closeModal} sx={{
+                                m: 5, maxWidth: 60, color: "#D05F5F", boxShadow: 0, borderRadius: '15px'
+                            }}><DoDisturbOnIcon/></Button>
+                        </Stack>
+                    </CardActions>
+                </Card>
             </Modal>
 
+            {/*<NewGameCard onClose={closeModal} loadCards={loadAllNonDefaultGameCards}/>*/}
 
             {/*<FormControl fullWidth>
                 <InputLabel id="demo-simple-select-label">Select game set</InputLabel>
@@ -89,7 +147,6 @@ export default function GameCardCollection() {
                 onGameCardChange={loadAllNonDefaultGameCards}
                 cardSetName={card.cardSetName}
                 title={card.title}
-                loadAll={loadAllNonDefaultGameCards}
             ></GameCardFrame>)}
 
         </>
