--- frontend/src/Collection/NewGameCard.css
@@ -1,9 +1,0 @@
-.new-card {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-#new-card-button {
-    margin: 0;
-}
